requirements.txt
fastapi==0.111.0
fastapi-mail==1.4.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
psycopg[binary,pool]==3.1.18
pyotp==2.9.0
pypng==0.20220715.0
pyqrcode==1.2.1
langchain==0.1.17
langchain-community==0.0.36
langchain-openai==0.1.3
langchain-postgres==0.0.11
celery[sqlalchemy]==5.3.6
boto3==1.34.106
pypdf==4.2.0
rapidocr-onnxruntime==1.3.17
langgraph==0.0.40
langchainhub==0.1.15
pytest==8.2.0
